requirements.txt
django==1.11
pillow
django-ckeditor
python-dateutil
psycopg2
